fix(hcl2cdk): remove circular dependency
