+++ packages/@cdktf/hcl2cdk/lib/expressions.ts
@@ -2,7 +2,6 @@ import * as t from "@babel/types";
 import reservedWords from "reserved-words";
 import { camelCase, pascalCase } from "./utils";
 import { TerraformResourceBlock, Scope } from "./types";
-import isValidDomain from "is-valid-domain";
 import { getResourceNamespace } from "@cdktf/provider-generator";
 
 export type Reference = {
@@ -445,24 +444,4 @@ export function findUsedReferences(
     (carry, i) => [...carry, ...findUsedReferences(nodeIds, i)],
     []
   );
-}
-
-// Logic from https://github.com/hashicorp/terraform/blob/e09b831f6ee35d37b11b8dcccd3a6d6f6db5e5ff/internal/addrs/module_source.go#L198
-export function isRegistryModule(source: string) {
-  const parts = source.split("/");
-  if (
-    source.startsWith(".") ||
-    parts.length < 3 ||
-    parts.length > 4 ||
-    source.includes("github.com") ||
-    source.includes("bitbucket.org")
-  ) {
-    return false;
-  }
-
-  if (parts.length === 4 && !isValidDomain(parts[0])) {
-    return false;
-  }
-
-  return true;
 }
