+++ packages/@cdktf/hcl2cdk/package.json
@@ -36,7 +36,6 @@
     "glob": "7.1.7",
     "graphology": "^0.20.0",
     "graphology-types": "^0.19.2",
-    "is-valid-domain": "^0.1.2",
     "jsii-rosetta": "^1.30.0",
     "prettier": "^2.3.1",
     "reserved-words": "^0.1.2",
