+++ packages/@cdktf/provider-generator/lib/index.ts
@@ -1,3 +1,4 @@
 export * from "./get/constructs-maker";
 export * as config from "./config";
 export { getResourceNamespace } from "./get/generator/constants/provider-namespaces";
+export { isRegistryModule } from "./get/module";
