+++ packages/@cdktf/provider-generator/package.json
@@ -30,6 +30,7 @@
     "cdktf": "0.0.0",
     "codemaker": "^0.22.0",
     "fs-extra": "^8.1.0",
+    "is-valid-domain": "^0.1.2",
     "jsii-srcmak": "^0.1.272"
   },
   "devDependencies": {
